+++ CHANGELOG.md
@@ -4,12 +4,13 @@
 
 - Added inverse lerp functions to support `float`, `float2`, and `float4` operations.
 - Added default fields to `VXEditorUtil`.
-
-### Changes
-
 - `FoldableReorderableList`:
   - Supports serialized structs/classes.
   - Element height now depends on serialized property height.
+  - Added "empty message" to display when the list is empty.
+
+### Changes
+
 - Removed `AbstractVXScriptableEditor`.
 - `serializedObject` now updates every frame in `OnInspectorGUI` (in `AbstractVXEditor`) to refresh `serializedObject` representation (especially in scriptable objects).
 - Removed unused functions from `VXEditorUtil`.